chore(api): Nettoyage cron inscription goal
